requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
pyarrow
tqdm
pypdf
